refactor: selection docs
