+++ src/player_interaction/selection/states.rs
@@ -1,0 +1,13 @@
+use bevy::prelude::*;
+
+#[derive(States, Debug, Clone, Eq, PartialEq, Hash)]
+pub enum CanPlaceState {
+    True,
+    False,
+}
+
+impl Default for CanPlaceState {
+    fn default() -> Self {
+        Self::True
+    }
+}
